Resolve import conflicts
