+++ neuromation/api/__init__.py
@@ -51,7 +51,6 @@ from .jobs import (
     JobStatusHistory,
     JobTelemetry,
     Resources,
-    SecretFile,
     StdStream,
     Volume,
 )
@@ -79,7 +78,6 @@ __all__ = (
     "Resources",
     "StdStream",
     "Volume",
-    "SecretFile",
     "HTTPPort",
     "Users",
     "Action",
